test(cli): cover default login provider
